refactor : Refactor StringAddCalculator
* Add StringConverter

* Delete the part to covert string to PositiveNumbers in StringAddCalculator
